messenger/requirements.txt
pylint~=2.15.3
coverage~=6.5.0
gunicorn~=20.1.0
Django~=4.1.2
psycopg2~=2.9.5
djangorestframework~=3.14.0
social_auth_app_django~=5.0.0
django_celery_results~=2.4.0
django-celery-beat~=2.4.0
beautifulsoup4~=4.11.1
celery~=5.2.7
flower~=1.2.0
bs4~=0.0.1
Pillow~=9.3.0
